Unpluralize Hasher namespace
