+++ src/Enums/Enum.php
@@ -7,8 +7,8 @@ use PHP\Collections\ByteArray;
 use PHP\Collections\Dictionary;
 use PHP\Enums\Exceptions\MalformedEnumException;
 use PHP\Hashing\HashAlgorithms\SHA256;
-use PHP\Hashing\Hashers\PrimitiveHasher;
-use PHP\Hashing\Hashers\SerializeAndHash;
+use PHP\Hashing\Hasher\PrimitiveHasher;
+use PHP\Hashing\Hasher\SerializeAndHash;
 use PHP\ObjectClass;
 use PHP\Serialization\PHPSerializer;
 use ReflectionClass;
