+++ src/Hashing/Hasher/HasherDecorator.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace PHP\Hashing\Hashers;
+namespace PHP\Hashing\Hasher;
 
 /**
  * Modifies hash results using the Decorator pattern
