+++ src/Hashing/Hasher/IHasher.php
@@ -1,7 +1,7 @@
 <?php
 declare( strict_types = 1 );
 
-namespace PHP\Hashing\Hashers;
+namespace PHP\Hashing\Hasher;
 
 use PHP\Collections\ByteArray;
