+++ src/Hashing/Hasher/PrimitiveHasher.php
@@ -1,15 +1,14 @@
 <?php
 declare(strict_types=1);
 
-namespace PHP\Hashing\Hashers;
+namespace PHP\Hashing\Hasher;
 
 use PHP\Collections\ByteArray;
-use PHP\Interfaces\IEquatable;
 
 /**
  * Hashes primitive value types (doubles, ints, strings) by simply returning their value as an array of bytes
  */
-class ReturnEquatableHash extends HasherDecorator
+class PrimitiveHasher extends HasherDecorator
 {
 
     /**
@@ -17,6 +16,8 @@ class ReturnEquatableHash extends HasherDecorator
      */
     public function hash($value): ByteArray
     {
-        return ($value instanceof IEquatable) ? $value->hash() : $this->getNextHasher()->hash($value);
+        return is_float($value) || is_int($value) || is_string($value)
+            ? new ByteArray($value)
+            : $this->getNextHasher()->hash($value);
     }
 }
