+++ src/Hashing/Hasher/ReturnEquatableHash.php
@@ -1,14 +1,15 @@
 <?php
 declare(strict_types=1);
 
-namespace PHP\Hashing\Hashers;
+namespace PHP\Hashing\Hasher;
 
 use PHP\Collections\ByteArray;
+use PHP\Interfaces\IEquatable;
 
 /**
  * Hashes primitive value types (doubles, ints, strings) by simply returning their value as an array of bytes
  */
-class PrimitiveHasher extends HasherDecorator
+class ReturnEquatableHash extends HasherDecorator
 {
 
     /**
@@ -16,8 +17,6 @@ class PrimitiveHasher extends HasherDecorator
      */
     public function hash($value): ByteArray
     {
-        return is_float($value) || is_int($value) || is_string($value)
-            ? new ByteArray($value)
-            : $this->getNextHasher()->hash($value);
+        return ($value instanceof IEquatable) ? $value->hash() : $this->getNextHasher()->hash($value);
     }
 }
