+++ src/Hashing/Hasher/SerializeAndHash.php
@@ -1,11 +1,11 @@
 <?php
 declare( strict_types = 1 );
 
-namespace PHP\Hashing\Hashers;
+namespace PHP\Hashing\Hasher;
 
 use PHP\Collections\ByteArray;
 use PHP\Hashing\HashAlgorithms\IHashAlgorithm;
-use PHP\Hashing\Hashers\IHasher;
+use PHP\Hashing\Hasher\IHasher;
 use PHP\Serialization\ISerializer;
 
 /**
