+++ tests/Hashing/Hasher/PrimitiveHasherTest.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace PHP\Tests\Hashing\Hashers;
+namespace PHP\Tests\Hashing\Hasher;
 
 use PHP\Collections\ByteArray;
-use PHP\Hashing\Hashers\HasherDecorator;
-use PHP\Hashing\Hashers\IHasher;
-use PHP\Hashing\Hashers\PrimitiveHasher;
+use PHP\Hashing\Hasher\HasherDecorator;
+use PHP\Hashing\Hasher\IHasher;
+use PHP\Hashing\Hasher\PrimitiveHasher;
 
 /**
  * Tests PrimitiveHasher
