+++ tests/Hashing/Hasher/SerializeAndHashTest.php
@@ -6,8 +6,8 @@ namespace PHP\Hashing\Hasher;
 use PHP\Collections\ByteArray;
 use PHP\Hashing\HashAlgorithms\IHashAlgorithm;
 use PHP\Hashing\HashAlgorithms\MD5;
-use PHP\Hashing\Hashers\IHasher;
-use PHP\Hashing\Hashers\SerializeAndHash;
+use PHP\Hashing\Hasher\IHasher;
+use PHP\Hashing\Hasher\SerializeAndHash;
 use PHP\Serialization\ISerializer;
 use PHP\Serialization\PHPSerializer;
 use PHPUnit\Framework\TestCase;
